Make methods static if possible
Summary:
Find methods that can be static, and make them so.  The heavy lifting
of this pass is outsourced to `devirtualize`, which determines whether a method
"needs to be" virtual, i.e., it is an interface implementation, or has
overrides elsewhere in its hierarchy.  Aside from that, the only check we need
is whether the method uses `this`.

Reviewed By: satnam6502

Differential Revision: D3948650

fbshipit-source-id: 7d5ab2666093f61fe79d7c0faf7abd4902adfbea