Make methods private if possible
Summary:
Find devirtualizable methods that are only called from their defining class,
and make them private.  `devirtualize` does the hierarchy analysis, so we just
need to scan for references from outside the defining class.

Note that we won't privatize methods that we have staticized earlier in this
pass.  I'm not sure whether invoke-direct or invoke-static is preferable, so I
made an arbitrary choice.  We could check this at some point with scanlab.

Reviewed By: satnam6502

Differential Revision: D3954750

fbshipit-source-id: 2fe1ee00d2ec045a75087548dd7b5b598724673c