Derive keeprules from new parser
Summary:
This is a temporary hack (hopefully not in the "Nothing's so permanent
as..." sense) to use the new parser to implement the functionality of the old.
In reachable-classes we're importing a subset of the proguard rules, so let's
synthesize that subset and apply them.  We should remove this once we've done a
thorough audit of how these rules are used, but in the meantime this will
unblock migration to the new parser.

The motivation is that I'd like to add a redex-only feature to the proguard
config language, and I don't want to maintain the lex/yacc parser.

Reviewed By: satnam6502

Differential Revision: D3997206

fbshipit-source-id: 3ae14752a2208fadc435efeba3eea3b247a687fa