Update stringly-typed method refs to account for colon
Summary:
We added a colon to the method descriptor string used by ProguardMap
and the canonical-name form of DexMethod::get_method.  This is why using strings
to represent complex data is bad.

This was causing remove-unreachable to go haywire, since it's driven by seeds
information, which is in turn parsed using these assumptions.

Reviewed By: satnam6502

Differential Revision: D4000864

fbshipit-source-id: 561b67794f518d534cce16dab6ee9580bfb44ff6