Stop using /tmp for metadata files
Summary:
We rewrite configs to avoid referencing /tmp for metadata (because
writing to `/tmp/hardcoded.name` is racy), and then move those metadata files
to the apk directory.  Let's just cut out the /tmp reference, and write
directly to the output dir.  (Of course, the output dir for redex-all is the
dex dir, so we still have to move it from there to the destination dir.  Sigh.)

Reviewed By: satnam6502

Differential Revision: D4000980

fbshipit-source-id: ad89dc51ce03f5e19127e55c4d25641d35a3f46c