Fix for array references in remove unreachable
Summary:
If a type is an array type we want to get to the "raw" type and move it from there.
The issue is that no array type are backed up by a DexClass and so the type_class resolution returns nullptr stopping the walk.
The unit test should clarify the issue.
It's rather unusual to have an array reference without having a reference to the naked/raw class anywhere in the code but it may happen

Reviewed By: onomatopoeia

Differential Revision: D4010792

fbshipit-source-id: b827a13856e6845a2cd2bd121f80465f3e46dc51