Output <meth/cls id, dex checksum> entries to symbol files
Summary:
The dex ordinal number (1st, 2nd, etc) is really hard to determine on-device.
We do have easier access to the dex header though, which means that we can use
the adler checksum to determine the identity of the dex.

This diff adds a the method or class id + dex checksum entries to the existing
symbol files, alongside the ordinal number entries. This is a
backward-compatible change (accepting the risk of a collision between an
ordinal and a checksum (<num dexes>/2^32). The ordinal entries will be removed
once the client has been migrated entirely.

I also deemed the change trivial and safe, so backported it to the stable redex
codebase as well.

build[automation_fbandroid_fallback_release]

Reviewed By: onomatopoeia

Differential Revision: D4016204

fbshipit-source-id: e4ce5345393d9f90e2281bd8eaed18dcc5cc699a