Fix proguard rename map merge
Summary:
The Python code that merges the Proguard and Redex class rename maps expects to
combine a Proguard-generated mapping of

  original name -> proguard-obfuscated name

and a Redex-generated mapping of

  original name -> redex-obfuscated name

into one big map. However, Redex was generating a mapping of

  proguard-obfuscated name -> redex-obfuscated name

as such, the only time we would create a correct map entry was if Proguard had
not obfuscated the name.

This diff fixes the redex binary to generate the

  original name -> redex-obfuscated name

mapping that the Python wrapper expects.

In order to make any future breakages more obvious, I've made the Python script
throw an error if it is unable to match a Redex map entry to a corresponding
Proguard map entry. For this to work I had to prune a number of irrelevant
entries from the Redex class map (e.g. array aliases), which explains the bulk
of the lines in this diff.

Reviewed By: onomatopoeia

Differential Revision: D4019174

fbshipit-source-id: 22f5fe427e5ac07ee1241113ce6eaf466682d6a5