Proguard Matcher adjustments
Summary:
Undo a previous "fix" which turned out to be wrong.
This should ensure that for "class * extends A" will
apply to the class matched by "*" and not just
its superclasses.

Also add a feature to dump out the class steucture
of the final scope (when printin of seeds is enabled).
This makes it easier to compare the incoming classes
with the outgoing classes.

Reviewed By: dariorussi

Differential Revision: D4051196

fbshipit-source-id: c784ad93b76fa25654e877ed60dffd20ec770797