simpleinline: Fix up debug info when making methods static
Summary:
The initial implementation of make_static neglected to update the debug info to
include the extra parameter. On ART[*], this resulted in the runtime failing to
parse the debug info. The practical consequence of this was that if the method
was part of any stack trace, there would be no file:line information available
for it.

[*] I did not check under Dalvik; I did check on N, and it doesn't seem to repro
there (though it could possibly have manifested under a different message)

Reviewed By: dariorussi

Differential Revision: D4050373

fbshipit-source-id: 9aac3fe359414ed2ac78bac4bf54876504c71a33