Allow specification of multiple ProGuard files
Summary:
This change to redex.py and redex-all allows multiple ProGuard configuration files to
be specified. This is needed so we can feed into redex the list of library and program
JARs from the upstream build process without relying on the single top level ProGuard
file (which will no longer be produced when ProGuard is removed).

Reviewed By: int3

Differential Revision: D4064558

fbshipit-source-id: 313e2e5615cfaf4e15dbe348097391b8bb960efb