Stop RenamerV2 from running and crashing when classes in hierarchy cannot be resolved
Summary:
right now few flavors of our build crash in redex (e.g. buck build fb4a-redex). That's because renamer v2 asserts if a type cannot be resolved to a class which on certain builds is a fact as they do not get external jars (see D4064958).
So this diff stops the renamer v2 pass from running if there are issues with resolution of base classes.
It also prints a warning which right now it's not displayed because we have warnings off. I think we need to improve our warning story and either have level or classes of warnings or both. But that is for another day. Once we have that we should have a basic level of warnings on so we can see certain errors.

Reviewed By: onomatopoeia, int3

Differential Revision: D4069642

fbshipit-source-id: fc692497d66b7368a3bdb1c8140a182a127f8a1b