Make Redex ProGuard config parser compute class level seeds with full parity w.r.t. ProGuard
Summary: This diff adds the final change needed to make the Redex config parser compute class level seeds to full level of parity with ProGuard at the level of classes (but not members yet). This allows us to use keep(class) and allowobfuscation(class) safely in our code for classes (but not fields and methods). The error reporting for methods and fields is currently suppressed until that case is fully supported.

Reviewed By: dariorussi

Differential Revision: D4070931

fbshipit-source-id: 28c0ef27a728784dec1601a75d25f6457cfb41fb