Fix a corner case handling in AccessMarking
Summary:
Fix a corner case handling in AccessMarking, where range_size becomes zero in invoke_range opcodes. Zero range_size in invoke_range is not well defined. This fix changes to non-range invoke if range_size is zero.

The code was also increasing the base register by one. If this +1 hits the limit of the method's register count, there could be a crash in other logics.

While running redex-stats-bot, we found a crash (assertion) in a pass: Liveness analysis during SimpleInline pass. The assertion was because of an attempt to set a 19-th register slot in a bitset while the method has only 18 registers.

The original instruction was: invoke-direct/range {v17}. But the AccessMarking pass converted to invoke-static/range {v18} (with range_size of 0). The method had only 18 register count: a typical out-of-bound assertion.

Reviewed By: dariorussi

Differential Revision: D4086496

fbshipit-source-id: 3d93d719889e97005592febc5783887b4b5be05c