Add very simple pass to mark apks which redex operated on.
Summary:
There should be an easy, defininitive way to tell if redex was run on a given
apk. Future extensions to this idea might add the ability for other passes to
write add info to be logged here, add the list of passes run, and/or add the
source revision that redex was built from.

Reviewed By: dariorussi

Differential Revision: D4121608

fbshipit-source-id: 4f3c17bc93a7327452ba3240ad4b2a8d12fffebe