Massive refactor
Summary:
Refactored out the visitor and most of the boilerplate classes, replacing them with lambdas. Refactor eliminated about 100 lines, I am not sure whether or not it made the code more readable.

Please give feedback if there is a stylistically nicer/more readable way to write out the function pointers/lambdas/templating or if there is a more readable/efficient/safe way to pass functions than std::function objects. I chose them because it becomes a little more obvious what the function that is being passed does, as well as enforcing type constraints on the function.

Reviewed By: dariorussi

Differential Revision: D4109113

fbshipit-source-id: 88a7376d2d80461cf5dafb1e8e94fcf0f83fe258