Factor out instr test BUCK rules
Summary:
More factoring of the BUCK rules so I can add more tests in a clean fashion.

Previously, redex_instr_test was only used for the AccessMarking and
RemoveUnreachable tests. I've extended it to cover the main test target as well.

To be honest, I don't quite understand why redex_instr_test runs PG twice with
potentially different configs. I guess it was useful somehow for
RemoveUnreachable's seeds?  For the main test target, it now runs PG twice with
the same config, which is a little redundant but should not be a functional
change.

Reviewed By: dariorussi

Differential Revision: D4106472

fbshipit-source-id: 6a95da4eb4e4ccef2268216918ff0bd434128535