Update can_delete to use keep and allowshrinking config bits
Summary:
Make can_delete use the ProGuard config bits computed by the new parser and matcher.
This says it is OK to delete a class, field or member if either it is marked for
`keep`, or even if it is marked for `keep` if the `allowshrinking` bit is set then
it is allowable to delete this class, field and member if you think it is safe to do
so e.g. it is not reachable from another seed.

Reviewed By: dariorussi

Differential Revision: D4156371

fbshipit-source-id: 1c41191430a4b4f77e229acf41d9dc61ec2bd425