Only check for leftover class strings for RenameClassesV2
Summary:
In {D4082925}, I made the Python wrapper script throw an exception if it could
not find a class mapping in the PG-generated mapping.txt to update. This was
supposed to act as a sanity check. However, this only really works with the V2
pass -- the V1 pass is still liable to generate mappings that don't correspond
to actual classes (e.g. array types).

This diff simply enables the check only for the V2 pass.

Reviewed By: dariorussi

Differential Revision: D4160814

fbshipit-source-id: 0383240def4f9663fc7a057724d9a3fbc99c0d5e