Re-enable annotation removal pass
Summary:
We are enabling AnnoKill again and removing only the BUILD annotations for now.
I amde few changes to the code, mostly "cosmetic" and some real ones.
I made an anonyomous namespace for the functions private to the file. Changed the TRACE level for some logs.
I also made a change to compute all known annotations to the app and to see which are referenced by code.
References by code are particualarly important because we definitely don't want to remove an annotation that has got live code looking for them. We have very few instances of such annotation classes though they end up being quite a number of them (~55K).
The all live references to annotation is, of course, a bit of a doomed story because we don't have code outisde of our app that we can look into and see what annotation that is possibly using.
Anyway, this is the first diff for the anno kill and it gets rid of over 100K annotations so it's all good.

build[automation_fbandroid_ics_,automation_fbandroid_art_]

Reviewed By: int3

Differential Revision: D4150280

fbshipit-source-id: dd22289ae95ed77b0b4c3bef3a2579bb69a1796f