Have compare_transform work on text file input as well
Summary:
Sometimes I prefer to run dexdump myself first, cut out the few methods that I
care about in the text output, then run the compare script on that.

I also made the script strip out the "line=123" part of the dump, as that is
frequently useless noise, especially with Redex's strategy of making the line
numbers a global index.

Reviewed By: dariorussi

Differential Revision: D4173816

fbshipit-source-id: 5370c88c8aa04721eb0b2a91e27df2f77906f03b