Re-enable anno kill
Summary:
cleaned up the code massively so we are sure we run only what we care. Fixed logs.
This is sort of a fresh start.
Only build and blacklisted anno allowed (keep_annos is the entry in the config).
It seems to do fine and keep what we need to.

Reviewed By: shanenay

Differential Revision: D4203529

fbshipit-source-id: 9a876408d4a8d77b88c5a07cb83cd6997d07eb6b