Remove work classes using CleanupFeature pass.
Summary:
In the previous pass, the di sget calls were not handled appropriately,
which caused the verifier to trigger. This changeset handles it.

The changeset also handles removal of packed-switch cases. Added some
utilities and MethodTransform methods for it.

The fb4a.config contains unsafesupercls, unsafeinterace and futureremovallist,
which are necessary for future messenger code removed. Currently, these can
be safely ignored.

This also fixes a minor issue where the first instruction is a move inst, which
was causing an assert on some of the debug builds.

Reviewed By: dariorussi

Differential Revision: D4211848

fbshipit-source-id: 58fea5d4fe30abb59f7352635aa0d997cea4a370