simpleinline: change_visibility should act on class of ref
Summary:
Previously, the simpleinline code would attempt to make all class refs in an
inlined callee public. However, for instructions that operate on methods, it
would first resolve the method to its def, then make the class containing
that def public. What was really needed is to make the class of the method
_ref_ public. E.g. in

```
public class Qux {
  private class Foo {
    protected void noninlinable() {
    }
  }

  private class Bar extends Foo {}

  Bar mBar = new Bar();

  public void inlineMe() {
    mBar.noninlinable();
  }
}
```

The method def is Foo.noninlinable(), but the method ref is Bar.noninlinable;
therefore, for inlineMe() to be inlined into a different class, Bar must be
made public.

Reviewed By: dariorussi

Differential Revision: D4204153

fbshipit-source-id: 4737c7692f2c4285383faf7ab9eea035a20fa7e7