Fix StaticRelo stats
Summary:
Previously, our TRACE output printed out the stats output
for the last DexStore processed, instead of for all Dexes. This fixes it.

num_moved_methods in redex-stats.txt was also wrong -- s_meth_move_count was
a cumulative counter, but we were adding it repeatedly to the PassManager's
cumulative counter, so we were double-counting a lot of things.

Reviewed By: dariorussi

Differential Revision: D4248898

fbshipit-source-id: c5d984ccfeed1e618d8dd875aa32a40a1b70a139