VMethod Renaming
Summary:
Changes included in the diff:
- DexElemManager now stores pointers so that we can subclass DexNameWrapper properly without object slicing
- MethodNameWrapper subclasses DexNameWrapper and adds the ability to "link" elements together. Linking elements means that their renaming information is linked so that all linked elements will be renamed at the same time (or all of them will be unrenamable).
- Essentially duplicated much of Devirtualizer logic in the Devirtualizer with slightly different data structures to do linking of vmethods correctly.

The way that vmethod renaming works is that the Devirtualizer walk figures out for every method what methods it needs to be "linked" to and links them. Overrides in a hierarchy are linked as well as implementations of interfaces. After linking, renaming is trivial because if we rename one of the linked members, the new name propogates to all of them. The devirtualizer also records information about what classes need to be walked to not have names conflict when implementing an interface. The devirtualizer also figures out what methods cannot be renamed by checking for classes or interfaces that are external and markin non-renamable methods.

build[automation_fbandroid_redexunstable,automation_messenger_redexunstable]

Reviewed By: dariorussi

Differential Revision: D4180541

fbshipit-source-id: dce135b3db0704af94e3347c5b1c9c2bf5d4236e