One more type of dex identifier in method/class symbols
Summary:
Our experiment with dex checksum worked great on art, where dex2oat preserves
the input checksum. Dalvik, however, recalculates it. Of course it does.

Instead of messing with dexopt to disable the recalculation (which is fairly
trivial but not open-source-able), let's use the SHA1 `signature` field, which
is //meant// for file identification, judging by the comments. We're not gonna
use the full thing though, just the top 32 bits read as an integer - ought to
be no worse than the adler32 `checksum` field.

Reviewed By: dariorussi

Differential Revision: D4302483

fbshipit-source-id: a556aeb803553a4b453f66bcf5de36616cbdd2ec