More pass stats instrumentation
Summary:
Several extra passes now will be integrated with StatsBot.  These are the easy ones to instrument, there are a few I left behind which I'll address in a separate diff.  Another 10 to go.

Changes are pretty trivial, the way this is done is not super idiomatic nor super consistent but gives us a quick way to get to our goal and I believe it doesn't look super ugly.

Reviewed By: newobj

Differential Revision: D4330962

fbshipit-source-id: f29334cdb17ef0274bdce66add298527d4bf066e