replace_opcode doesn't work for branch opcodes
Summary:
replace_opcode currently checks if the "from" opcode is a branch opcode, and if
so, removes any BranchTargets that are pointing to it. This is the right
behavior if the "to" opcode is *not* a branch opcode, since we don't want any
dangling BranchTarget pointers. However, ConstProp is substituting one branch
opcode for another, and it indeed wants the new branch opcode to jump to the
same location as the old one. So we *don't* want to delete the BranchTarget in
that case. To minimize confusion, I've created a new method call --
replace_branch -- that specifically handles that case.

In the longer run we probably want some sort of API that allows us to relocate
BranchTargets as well, instead of our currently fairly limited options of
preserving or deleting the existing one.

Reviewed By: dariorussi

Differential Revision: D4335745

fbshipit-source-id: 6bd0ab17f48fa43c44c87572a037214fde295b9e