Include method name in line number map
Summary:
The initial version of the line number map only included the
file name and line number, but lacked method names. It
didn't matter too much back then as we didn't obfuscate
method names. However, now that we are going to do so,
the line number map needs to include this information.

(Emitting a PG map with this information was also an option,
but I think this format is easier to work with)

For now, redex will emit the maps in both the old and the
new formats to ease the migration process.

Reviewed By: dariorussi

Differential Revision: D4324225

fbshipit-source-id: 2120534512231bb62d3832207fa7558c048dee69