Implement Peephole V2 optimization with several string replacements
Summary:
Implement a similar optimization to PG's code/simplification/string.
This is a peephole optimization with a huge number of code patterns.
As our initial peephole optimization is not ideal to handle such many
patterns, I implemented a simple pattern-based peephole optimizer.

TODO: there are 100+ patterns in PG. But most of them are variations of
different types. The first commit has a few of them. I will add more.

Reviewed By: satnam6502

Differential Revision: D4318930

fbshipit-source-id: b61d198dc66dd82e3fcad8266e9dd509609d9a87