Produce a single-file self-extracting script for OSS
Summary:
For the benefit of Buck, homebrew, and all the users who have no idea
what redex-all is, we'll now build redex as a magic self-extracting script.  The
script contains a bash extraction preamble followed by a tar file containing
redex-all and the python wrapper (and libraries).  When `redex` is invoked it
untars this payload to a temp location, executes redex with the specified
parameters, and then deletes the untarred payload.

Differential Revision: D4401602

fbshipit-source-id: 94fa04d7a966be55ab86720c4d426ea566b562bb