Refactor MethodTransform's representation of fill-array opcodes
Summary:
A pseudo-opcode that is "jumped" to by the fill-array opcode is awkward to work
with. Practical consequence is that the inliner currently doesn't support
inlining callees with this instruction, and LocalDCE works around it pretty
awkwardly with its can_delete check. This diff stores the opcode in a separate
hashtable instead.

I'll have the inliner support this opcode in a follow-up diff.

Reviewed By: dariorussi, bertmaher

Differential Revision: D4397080

fbshipit-source-id: 9da92e8ba9aa0effa74adee51667acf15372318f