reenable const prop, with return prop disabled
Summary:
Return propagation in const prop seems like a hopelessly broken idea. It regards any method with a known return value to be completely elidable, even if the method has side effects. This will require a pretty serious overhaul to get working, and I'm not even sure what the possible impact is. Even as written, it was only propagating 22 (!) returns!

It's a pretty soft disable. Max and I will probably dig into this more, and if we determine that the opportunity really is nill, I will come back and delete the code more thoroughly. If there's an opportunity, we'll move forward.

Reviewed By: bertmaher

Differential Revision: D4415615

fbshipit-source-id: 46e4226adc71eb94e50e8a634efb3f30e839c8e5