handling-multi-byte-dex-characters-in-peephole
Summary:
This is a first iteration of handling multiple byte dex mutf8 char in peephole V2.

1. Add function size_of_mutf8_char that calculates the byte size of a mutf8 char
2. Add function mutf8_char_to_cstring that shuffles a mutf8 char literal int to a cstring
3. Updated matching rules and logic to handle multiple byte dex mutf8 char literal
4. Updated simplify_string_verify test to account for newly enabled savings

Couple of things remaining.
It doesn't quite handle \u0000 for mutf8 at the moment.
I added a \u0000 test case in test_Replace_ValueOfChar. But somehow it cannot be matched by the opt rule..

It is not entirely clear to me how DexString is handled.
Are they all converted from mutf8 to regular utf8 used in std::string? and converted back when writing out the optimized apk?

Reviewed By: minjang

Differential Revision: D4461305

fbshipit-source-id: e9b8676ddff1b189e16181561fa861cee843022d