Add PrintMembers pass as a utility
Summary:
It's really useful at times to just be able to dump a concise list of
what's in a dex (classes, fields, methods).  `dexdump` does this but the output
is verbose, and I'm tired of writing grep/sed/awk/perl commands to munge it when
we can just print stuff the right way easily.

Reviewed By: satnam6502

Differential Revision: D4476772

fbshipit-source-id: ea59a0632253c4cef143c84366771b4176ee91f8