Have CleanupFeature rely on LocalDCE to clean up dead code
Summary:
No functional change, but this simplifies the remove_switch_case code a fair
bit. Instead of walking the MIE list to figure out the start and end of a case block,
we can just remove the particular MFLOW_TARGET that is the label of the case
block, which causes the instructions in that block to be recognized by localDCE
as dead code.

I'm trying to refactor Transform.cpp and so I'm trying to minimize the number of
places that deal with the guts of FatMethod

Reviewed By: dariorussi

Differential Revision: D4438395

fbshipit-source-id: f3af33784be303d9fe085f6a2c3f33e376d2484a