Support conditional marking of class members
Summary:
This diff introduces conditional marking for class members, which
essentially means, "If the containing class is a root, then this member is a
root also."  This is useful in two places:

1. keepclassmembers

   Proguard semantics require keeping these members only if the class itself is
   kept.

2. Implementations of interface methods.

   If a method implements a kept interface method, we only need to keep it if
   the containing class is kept.  (E.g., consider a class implementing List that
   is never instantiated; no need to keep it just because it implements the List
   interface.)

A conditionally marked member will be visited immediately if its containing
class is already marked; otherwise, it will be visited when its containing class
is visited.

Also, this gets rid of the automatic visiting of <init>.  I'm pretty sure that
was just paranoia talking.

Reviewed By: satnam6502, dalves

Differential Revision: D4440870

fbshipit-source-id: e12f5b5b2fd2b82cfdec7a07aa6ea18bbe4b4b1f