Don't insert MFLOW_FALLTHROUGH at the start of blocks when building CFG
Summary:
They were initially used to make it possible to delete opcodes without
invalidating the FatMethod iterators that indicated the start / end markers of
a basic block. However, I don't think this is necessary any more. When we
delete opcodes, we now zero them out (by setting them to MFLOW_FALLTHROUGH),
instead of removing the MethodItemEntry from the list. As such we do not have
to worry about keeping MIEs around to indicate the start / end positions of the
basic block boundaries.

Reviewed By: dariorussi

Differential Revision: D4542257

fbshipit-source-id: a1276c0d557698386abff717ec625e4420a923a3