Merge AnnoKill and AnnoClassKill into a single pass
Summary:
The next installment on revisiting annotation optimizations in Redex!  In this case, `AnnoClassKill` can be merged into `AnnoKill` and possibly even augment its capability: `AnnoClassKill` receives a set of annotations that we know for sure can be killed, and then expands this list of types by adding all annotation types that are in turn annotated with this set.  The original `AnnoClassKill` pass only removes this new set of annotations from method signatures, by combining `AnnoKill` and `AnnoClassKill` now it will effectively remove this set of annotation from anywhere (Classes, Methods and Fields).

Kind of risky, this is highly experimental don't try at home type of thing.

Reviewed By: dariorussi

Differential Revision: D4524917

fbshipit-source-id: f0848a3af6383557566471d2119612a8196b83b5