Remove unused try regions
Summary:
In addition to taking up space, leaving unused try regions around could
result in verification errors. For example, supposed we had a method with
code like

  TRY_START
  <some non-throwing code like const/4...>
  TRY_END
  CATCH_START
  move-exception ...

LocalDCE would see that the code between TRY_START ... TRY_END would not throw,
and therefore the code following CATCH_START was dead code. It would delete
that code but neglected to remove the TRY entries pointing to it. As a result,
we would emit a try region that pointed to the CATCH_START which is now past
the end of the method (since the opcodes following it have been deleted).

Reviewed By: minjang

Differential Revision: D4564484

fbshipit-source-id: 245c9c233a02498a75d72fe349c520e88be973b7