Validate keystore is present when `--sign` is used
Summary:
When default keystore is not present and `--sign` is passed to
Redex you'll get a type error when `jarsign` is called because `None`
values are provided for keystore and friends.  With this change the code
will exit early if keystore is not provided as an argument and the
default value turns out to be `None`.
Closes https://github.com/facebook/redex/pull/207

Differential Revision: D4577660

Pulled By: bertmaher

fbshipit-source-id: af3d365d8a313f3206015a7945e5f5f52806f778