Make dataflow analysis functions available in header
Summary:
This is a pure refactor w/ no functional change.

backwards_dataflow is currently only being used by RegAlloc, but I can see
it gaining more use cases.

In addition to this factoring, I've also made the transfer function a separate
argument to the dataflow function, instead of an instance method on the lattice
element. This makes it possible for the transfer function to store some state,
which will be useful in an upcoming diff.

Reviewed By: newobj

Differential Revision: D4578186

fbshipit-source-id: 4b355d742dd5a7b5028c31969f808b035ce5fae9