Fix literal set/get for arith-immediate with corrected handling of 21h
Summary:
The literal field for arithemetic opcodes like mul-int/lit8, and
also high opcodes like const/high16, was not matching the Dalvik spec.
Correct the handling of these literal fields and add a robust
test case for literals handling.

The previous attempt (D4535670) corrected form 21h for wide literals,
but broke it for narrow literals. In fact form 21h is annoying in that
the literal value depends on the opcode. Add a check for the opcode in
the literal getter/setter to distinguish these cases.

Reviewed By: bertmaher

Differential Revision: D4561992

fbshipit-source-id: ba67ecaedb396c4f8f56d7832adf148a89da69be