Account for java.lang.Object not to have a DexClass
Summary:
java.lang.Object does not have a DexClass when no external references are loaded.
The code was not accounting for that and it was crashing every time the virtual scope for Object was reached and there were no external references.

I was considering to always create a DexClass for java.lang.Object and I think I will do that in a later diff unless someone has got objections or concerns about it.

I wrote a unit test but that creates a merge problem with the renamer diff that is still up so if you guys are ok with it I'll add unit tests in a later diff

Differential Revision: D4605198

fbshipit-source-id: de1f735a8e781b28f871c80f8b379f6b23b4432c