Add arithmetic peephole optimizations and testing infrastructure
Summary:
Introduce some very simple peephole optimizations for arithmetic,
including multiplying or dividing by 1 and by -1, and also adding 0.

As part of this change, we equip Pattern with a predicate function.
The predicate runs if and only if the instructions match the pattern,
and allows the pattern to perform arbitrary logic (in this case, check
if literal values are matched).

We also introduce a robust peephole test infrastructure using DexAsm,
which allows for direct verification of peephole replacements.

This is the first peephole optimization in which the replacement
instructions have narrower register field widths than the replaced
instructions. Rather than make every pattern handle this possibility,
we introduce some logic in Pattern to compute the maximum field widths
for each pattern register, and and refuse to match register values that are
too wide.

Reviewed By: thezhangwei

Differential Revision: D4545699

fbshipit-source-id: 99d21fd36467429971871ba9ec66d117a0efe940