simpleinline: Fix "aliased inputs" check
Summary:
We do have a check to avoid inlining cases where the invoke instruction uses
the same register for multiple arguments to the callee, and the callee's body
overwrites at least one of their argument registers. This is because our simple
register remapping scheme won't handle these situations correctly.  However,
since the code for this check ran after `remap_reg_set`, it was possible for
`def_ins` -- the bitvector that tracks registers that were written to -- to be
incorrectly remapped. As a result the test would incorrectly pass and the
inliner would generate incorrect code.

Reviewed By: satnam6502

Differential Revision: D4626750

fbshipit-source-id: 6e3b4a02671bdd5b2d56d4537392401e853f83d3