Make CustomSort stricter
Summary:
In a separate diff, I attempted to use std::stable_sort()
instead of std::sort() for sorting code items. This caused APKs to
change unexpectedly, which led me to discover that CustomSort is
not nearly as strict as it could be.

Previously, the relative ordering of methods within a single class
was an artifact of the input data and the particular std::sort()
implementation we are using. After this change, the ordering should
be fully specified.

There's no real direct benefit to this change, but it will make it
less likely that other changes to this code produce APK changes,
which will make those changes easier to land. Further, it takes
any guesswork out of the relative ordering of methods within a class.

This diff will change APKs, and hence could theoretically cause a
regression, but in that case, the fact that the previous ordering
was "better" was pure luck, as we never specified that ordering
anywhere.

Reviewed By: dariorussi

Differential Revision: D4629458

fbshipit-source-id: a768132fe6223bc52355c857e8e83a0e99d5fd3f