Fix virtual scope escaping on missing interface
Summary:
Escaping was not correct in all cases.
Consider the following

  class A { void m() {} }
  class B extends A implements I { void f() {} }
  class C extends B { void g() {} }
  class D extends A { void h() {} }

and say I is unknown.
In that case we want all the methods down the hierarchy from B to be ESCAPED and all of those up the hierarchy to be escaped as well, but not those introduced in any sibling of B.
We were not marking all the methods down the hierarchy from B as ESCAPED.
So effectively in that case only D.h() should not be ESCAPED but instead we did not mark C.g() as well.

The specific failure was the following

  abstract class UnmodifiableListIterator implements ListIterator {
    // several methods but no implementation of ListIterator.hasNext()
  }

  class SomeClass extends UnmodifiableListIterator {
    bool hasNext() { ... }
  }

and ListIterator is unknown in the perf build (missing external jars).
The code was not going down the hierarchy and marking everything ESCAPED and thus left SomeClass.hasNext() FINAL|TOP_DEF thus making it elegible for inlining and removal. Which we did...

Reviewed By: int3

Differential Revision: D4654612

fbshipit-source-id: 4db7c9a005e60c6ada0280ecebd7847caa9d97d3