Fail fast if Redex config has wrong JSON
Summary:
Every time I want to run redex with a hand crafted config I have to wait 20secs for the tool to vomit in some mysterious way sometimes.  If I have a wrong `.config` file, please don't waste my time unpacking the `.apk`.

Also, question for you all, I'm tackling the task of why redex is so slow if we pass an empty config and the reason is unpack + repack.  Should we shortcut the case where `passes` is empty and do nothing?  I see some comments on the python script that says the packing algorithm use is better than the 'stock' one.

Reviewed By: dariorussi

Differential Revision: D4671430

fbshipit-source-id: abe4509ebe122f69e4268739bfb530ea1f13fbda