simpleinline: Change how we calculate method size threshold
Summary:
{D4151785} added this method size check -- see its commit message for more
context.

I am in the process of moving all instruction analysis to operate over
FatMethods instead of a vector of DexInstruction. One consequence of this is
that pseudo-opcodes will not be available. The method size check currently
includes these opcodes in its size estimation. Fortunately, this is not really
an issue, because ART's limit of 1 << 16 instruction units only cares about the
offset of the last "code point", and pseudo-ops are not code points. As such,
I've excluded the pseudo-ops from our method size estimates.

Reviewed By: dariorussi

Differential Revision: D4661419

fbshipit-source-id: dbf7e2d740152c27cd2730bf5d1a4ce1304e10c2