Simplest builder case
Summary:
First pass through builders and optimizing the most trivial ones.
For this case, I chose only the builders that:
* don't escape the stack
* they don't pass "this" in the constructor
* build method can be inlined
* all fields are initialized
After inlining the build method, if it succeeds, it tries to remove the builder completely. For the moment, the solution doesn't allocate extra registers. If possible, it removes the fields and reuse the registers.

Given that in fb4a there are none that follow this pattern, didn't add the pass. Following iterations will do that.

Reviewed By: int3

Differential Revision: D4674949

fbshipit-source-id: 532d77151b124a59d4ae23013d95b220292d331b